7_Memory_Search_API/src/requirements.txt
locust>=2.17.0
requests>=2.31.0
python-dotenv>=1.0.0



